==> GPCRdb_VEPs/1-1_GPCRdb_extract_mutation_EnsemblVEPinputs.ipynb <==
# README  
This notebook mainly did two things:  
Given the 400 olfactory-related GPCRs, extract their mutation data from GPCRdb. - get in vitro data  
At the same time, run Ensembl VEPs on the mutation data. - get VEP scores

# preliminary 


import pandas as pd
import os
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import re
import requests




# 20 types of amino acids
AA_all = {
    "A" : "Ala",
    "R" : "Arg",
    "N" : "Asn",
    "D" : "Asp",
    "C" : "Cys",
    "E" : "Glu",
    "Q" : "Gln",
    "G" : "Gly",
    "H" : "His",
    "I" : "Ile",
    "L" : "Leu",
    "K" : "Lys",
    "M" : "Met",
    "F" : "Phe",
    "P" : "Pro",
    "S" : "Ser",
    "T" : "Thr",
    "W" : "Trp",
    "Y" : "Tyr",
    "V" : "Val"
}

# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
# gene_list_raw
protein_list = gene_list_raw["protein"].tolist()
enst_list = gene_list_raw["ENST"].tolist()


gene_list_raw

# gname = gene_list[124]
# # gname = 'o2t11_human'
# # enst = enst_list [0]
# url = 'https://gpcrdb.org/services/mutants/'+gname
# response = requests.get(url)
# mutants_data = response.json()
# df = pd.DataFrame(mutants_data)
# df

# check how many proteins have mutants info 
count = 0
mutant_genes =[]
for i in range(len(protein_list)):
    gname = protein_list[i]
    url = 'https://gpcrdb.org/services/mutants/'+gname
    response = requests.get(url)
    mutants_data = response.json()
    if(len(mutants_data)>0):
      count = count + 1
      mutant_genes.append(gname)
    else:
      print(gname+' info not found!')

print('there are '+str(len(mutant_genes))+' genes have mutants info')

there are 166 genes have mutants info

# only extract the info of those 167 genes
mutant_genes_df = gene_list_raw[gene_list_raw['protein'].isin(mutant_genes)]
mutant_genes_df.reset_index(drop=True, inplace=True)
mutant_genes_df

# in 166 genes, find if their enst have info in ensemble(because later we need to use enst id as input to ensembl vep)
server = "https://rest.ensembl.org"
for i in range(mutant_genes_df.shape[0]):
    gname = mutant_genes_df['gene'][i]
    enst = mutant_genes_df['ENST'][i]
    # ensg = mutant_genes_df['ENSG'][i]
    ext = "/sequence/id/"+enst+"?type=cdna"
    r = requests.get(server+ext, headers={ "Content-Type" : "text/x-fasta"})
    if not r.ok:
      print(gname+' ENST not OK')

# Get access to GPCRdb

all ENSTs are OK! Now next extract mutation data. create two dataframes: gpcr_all_mutants stores all data from GPCRsb, vep_input all stores the mutation info(but transform from 1 letter abbreviation to 3 letter abbreviation) as the further inputs to Ensembl VEP

gpcr_all_mutants = pd.DataFrame()
vep_input_all = pd.DataFrame()
for i in range(mutant_genes_df.shape[0]):
  gname = mutant_genes_df['protein'][i]
  enst = mutant_genes_df['ENST'][i]
  url = 'https://gpcrdb.org/services/mutants/'+gname
  response = requests.get(url)
  mutants_data = response.json()
  df = pd.DataFrame(mutants_data)
  # only keep missense mutation
  df = df[df['mutation_to']!='-']
  df = df[df['mutation_to']!='X']
  df = df[df['mutation_to']!='']
  df = df[df['mutation_from']!=df['mutation_to']]
  df.reset_index(drop=True, inplace=True)
  gpcr_all_mutants = gpcr_all_mutants.append(df)
  vep_input_df = pd.DataFrame()
  print("processing No."+str(i)+" ---> "+gname)
  for j in range(df.shape[0]):
    row = vep_input_df.shape[0]
    vep_input_df.loc[row,'protein'] = gname
    vep_input_df.loc[row,'ENST'] = enst
    vep_input_df.loc[row,'mutation_pos'] = df.loc[j,"mutation_pos"]
    # we need to store this column so that we can filter out rows with no exp_fold_change value
    vep_input_df.loc[row,'exp_fold_change'] = df.loc[j,"exp_fold_change"]
    vep_input_df.loc[row,'mutation_from'] = AA_all[df.loc[j,"mutation_from"]] 
    vep_input_df.loc[row,'mutation_to'] = AA_all[df.loc[j,"mutation_to"]]
  vep_input_df['mutation_pos'] = vep_input_df['mutation_pos'].astype(int) 
  vep_input_df['vep_input'] = enst + ':p.' + vep_input_df['mutation_from'] + vep_input_df['mutation_pos'].astype(str) + vep_input_df['mutation_to']
  
  vep_input_all = vep_input_all.append(vep_input_df)



print(vep_input_all.shape)
print(gpcr_all_mutants.shape)

# we need to delete row with exp_fold_change==0 for both vep_input_all and gpcr_all_mutants
# drop the lines with exp_fold_change==0 
gpcr_filtered_mutants = gpcr_all_mutants[gpcr_all_mutants['exp_fold_change']!=0]
gpcr_filtered_mutants

unq_gpcr = gpcr_filtered_mutants.drop_duplicates(subset=['protein','mutation_pos','mutation_from','mutation_to'], keep='first', inplace=False)
unq_gpcr

vep_filtered_all = vep_input_all[vep_input_all['exp_fold_change']!=0]
vep_filtered_all

ll_vep = pd.DataFrame(np.unique(vep_filtered_all['vep_input'].tolist()))
#drop 'nan
# ll_vep = ll_vep[ll_vep[0]!='nan']
ll_vep

# output data

vep_filtered_all.to_csv('vep_filtered_all.csv',index=False)

# save the filtered vep inputs
ll_vep.to_csv('all_vep_inputs.csv',index=False, header=False)


# save to csv
gpcr_filtered_mutants.to_csv("gpcr_filtered_mutants_0217.csv")
