==> GPCRdb_VEPs/2-1_from_GPCRdb_to_Dynamut2.ipynb <==
# submit GPCRdb mutations to Dynamut2, get and process the results, then output to a csv file

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.metrics import r2_score
import re
import requests
import json
import os


gpcr_valid_missense = pd.read_csv('gpcr_valid_missense.csv')
gpcr_valid_missense

## Get AF structure from Uniprot

# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
uni_list = gene_list_raw['Uniprot'].tolist()
gene_list_raw

# a function to get the alphafold structure (pdb file) for each gene
def getAF(uniprot_id):
    # Set the API endpoint and the UniProt ID of the protein
    endpoint = "https://alphafold.ebi.ac.uk/files"

    # Send a GET request to the API to retrieve the protein's structure
    response = requests.get(f"{endpoint}/AF-{uniprot_id}-F1-model_v4.pdb")

    # Check the status code of the response to ensure it was successful
    if response.status_code == 200:
        # Write the contents of the response to a PDB file
        with open(f"AF_pdb/{uniprot_id}.pdb", "w") as f:
            f.write(response.text)
    else:
        print(f"Failed to retrieve structure for {uniprot_id}: {response.status_code}")



#get the structure
for i in range(len(uni_list)):
    getAF(uni_list[i])

## Make mutation list

# make the mutation file (needed input for Dynamut2) for each protein
uniid_list = []
gene_list = []
for i in range(gpcr_valid_missense.shape[0]):
    # get the uniprot id according to entryname and convert the Series to a string with index excluded
    uniprot_id = gene_list_raw.loc[gene_list_raw['protein']==gpcr_valid_missense.iloc[i,4],'Uniprot'].to_string(index=False)
    gene = gene_list_raw.loc[gene_list_raw['protein']==gpcr_valid_missense.iloc[i,4],'gene'].to_string(index=False)
    uniid_list.append(uniprot_id)
    gene_list.append(gene)
    mut_info = gpcr_valid_missense.iloc[i,30]+str(gpcr_valid_missense.iloc[i,1])+gpcr_valid_missense.iloc[i,31]
    with open(f"MutationLists/{uniprot_id}.txt", 'a') as file:
        # write a new line to the file
        file.write("A "+mut_info+"\n")
    

len(np.unique(gene_list))

unq_uniid_list = np.unique(uniid_list)
unq_uniid_list

# generate the bash file
for i in range(len(unq_uniid_list)):
        uniprot_id = unq_uniid_list[i]
        with open("curl_post.sh", 'a') as curl:
                # write a new line to the file
                curl.write(f"curl https://biosig.lab.uq.edu.au/dynamut2/api/prediction_list -X POST -i -F pdb_file=@/Users/joannahench/Documents/UCPH/MSc_Thesis/GPCRdb_VEPs/AF_pdb/{uniprot_id}.pdb -F mutations_list=@/Users/joannahench/Documents/UCPH/MSc_Thesis/GPCRdb_VEPs/MutationLists/{uniprot_id}.txt >>out.txt"+"\n")    

## process the output file to get the job id, also handle other issues

# if runs successfully, we store the job id, if not successfullt(if the mutation position not match the pdb structure), record the protein index in the protein list
import json
job_ids = []
block_ids = []

with open('out.txt') as f:
    block_idx = 0
    for line in f:
        if line.startswith('{'):
            data = json.loads(line.strip())
            # if successful, store the job id for further getting the results
            if 'job_id' in data:
                # print(f"Job ID: {data['job_id']}")
                job_ids.append(data['job_id'])
            # if met error, store the index for further delete the protein(these protein will not be used in the further analysis)
            elif 'error' in data:
                # print(f"Error: {data['error']}")
                # print(f"Block Index: {block_idx}")
                block_ids.append(block_idx)
            block_idx += 1

block_ids

# Remove values accordding to indices
ok_uniid = np.delete(unq_uniid_list, block_ids)
ok_uniid

# generate a df showing the uniprot id and the job id
uni_job_id = pd.DataFrame({"Uniprot":ok_uniid,"Job_id":job_ids})
uni_job_id

# according to job_ids, generate the bash file curl_get.sh
for i in range(len(job_ids)):
        job_id = job_ids[i]
        with open("curl_get.sh", 'a') as curl:
                # write a new line to the file
                curl.write(f"curl https://biosig.lab.uq.edu.au/dynamut2/api/prediction_list -X GET -F job_id={job_id} >>res0328.txt"+"\n")    

## get the dynamut2 prediction by running curl_get.sh 

## processing prediction

id = 0
data_df = pd.DataFrame()
with open('res0328.txt', 'r') as f:
    for line in f:
        data = json.loads(line.strip())
        df = pd.DataFrame(data).transpose().iloc[:-1]
        df[['mutation_from', 'mutation_pos', 'mutation_to']] = df['mutation'].str.extract(r'(\D)(\d+)(\D)')
        df['Uniprot'] = ok_uniid[id] 
        data_df = pd.concat([data_df, df], ignore_index=True) 
        id = id + 1



# define the new column order
new_col_order = ['mutation_from', 'mutation_pos', 'mutation_to','mutation','chain','prediction','Uniprot']
# reindex the columns with the new order
data_df = data_df.reindex(columns=new_col_order)
# merge with protein name
data_df = pd.merge(data_df,gene_list_raw.iloc[:,3:],on='Uniprot')
data_df

# output the results into a csv file
data_df.to_csv("dynamut2_out.csv")
