==> GPCRdb_VEPs/2-2_Dynamut2_ddg_GPCRdb_foldchange.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.metrics import r2_score
import re
import requests
import scipy as sp


# GPCRdb data
GPCRdb_fil = pd.read_csv("gpcr_filtered_mutants_0217.csv")
GPCRdb_fil


#Dynamut2 prediction
dynamut_pred = pd.read_csv("dynamut2_out.csv")
# rename the column EntryName for merging
dynamut_pred = dynamut_pred.rename(columns={'prediction':'Dynamut2'})
dynamut_pred


# merge two dataframes
gpcr_dyna  = pd.merge(GPCRdb_fil,dynamut_pred,on=['mutation_from','mutation_pos','mutation_to','protein'])
gpcr_dyna

gpcr_dyna['abs_fc'] = abs(gpcr_dyna['exp_fold_change'])
gpcr_dyna

gpcr_dyna.to_csv('GPCR_Dynamut2_df.csv',index=False)

gpcr_dyna = pd.read_csv("GPCR_Dynamut2_df.csv")
gpcr_dyna

value = 'exp_fold_change'
value1 = 'abs_fc'
vep = 'Dynamut2'

p_df1 = gpcr_dyna.groupby(by=["protein","mutation_pos","mutation_from"]).max()[[vep,value]]
p_df2 = gpcr_dyna.groupby(by=["protein","mutation_pos","mutation_from"]).min()[[vep,value]]
p_df3 = gpcr_dyna.groupby(by=["protein","mutation_pos","mutation_from"]).mean()[[vep,value1]]
#dataframe only include positive exp_fc
p_df_pos = gpcr_dyna[gpcr_dyna[value]>0]
#dataframe only include negative exp_fc
p_df_neg = gpcr_dyna[gpcr_dyna[value]<0]

# average of positive exp_fc
p_df4 = p_df_pos.groupby(by=["protein","mutation_pos","mutation_from"]).mean()[[vep,value]] 
# average of negative exp_fc
p_df5 = p_df_neg.groupby(by=["protein","mutation_pos","mutation_from"]).mean()[[vep,value]]

print("max(fc) with ddg:")
print(sp.stats.spearmanr(p_df1[value], p_df1[vep]))
print("\n")
print("min(fc) with ddg:")
print(sp.stats.spearmanr(p_df2[value], p_df2[vep]))
print("\n")
print("mean(abs(fc)) with ddg:")
print(sp.stats.spearmanr(p_df3[value1], p_df3[vep]))
print("\n")

print("mean(posfc) with ddg:")
print(sp.stats.spearmanr(p_df4[value], p_df4[vep]))
print("\n")
print("mean(negfc) with ddg:")
print(sp.stats.spearmanr(p_df5[value], p_df5[vep]))


## try another option: classify into three groups---accuracy, sensitivity, specificity

from sklearn.metrics import accuracy_score, recall_score, precision_score, f1_score, classification_report, roc_curve, auc, confusion_matrix

# ddg into 3 classifications: (neg infinity,-1), [-1,1],(1,pos infinity)
gpcr_dyna['Dynamut2_group'] = gpcr_dyna['Dynamut2'].apply(lambda x: 1 if x < -1 else (1 if x > 1 else 0))
gpcr_dyna['foldchange_group'] = gpcr_dyna['exp_fold_change'].apply(lambda x: 1 if x < -5 else (1 if x > 5 else 0))


true_labels = gpcr_dyna['foldchange_group']
predicted_labels = gpcr_dyna['Dynamut2_group']



# Calculate confusion matrix
tn, fp, fn, tp = confusion_matrix(true_labels, predicted_labels).ravel()

# Calculate accuracy
accuracy = accuracy_score(true_labels, predicted_labels)

# Calculate specificity and sensitivity
specificity = tn / (tn + fp)
sensitivity = tp / (tp + fn)


# Calculate precision, recall and F1-score
precision = precision_score(true_labels, predicted_labels)
recall = recall_score(true_labels, predicted_labels)
f1 = f1_score(true_labels, predicted_labels)

# calculate FPR and TPR
fpr, tpr, thresholds = roc_curve(true_labels, predicted_labels)

# calculate AUCROC
auc_roc = auc(fpr, tpr)



print("Specificity:", specificity)
print("Sensitivity:", sensitivity)
print("Accuracy:", accuracy)
print("Precision:", precision)
print("F1-score:", f1)
print("AUCROC score: {:.3f}".format(auc_roc))

# Generate classification report
report = classification_report(true_labels, predicted_labels, labels=[0, 1], digits=4,zero_division="warn")

# Print report
print(report)



# calculate FPR and TPR
fpr, tpr, thresholds = roc_curve(true_labels, predicted_labels)

# calculate AUCROC
auc_roc = auc(fpr, tpr)

# print the result
print("AUCROC score: {:.3f}".format(auc_roc))

accuracy = accuracy_score(true_labels, predicted_labels)
class_1_accuracy = accuracy_score([1 if label == 1 else 0 for label in true_labels], 
                                  [1 if label == 1 else 0 for label in predicted_labels])
class_0_accuracy = accuracy_score([1 if label == 0 else 0 for label in true_labels], 
                                  [1 if label == 0 else 0 for label in predicted_labels])
# class_minus_1_accuracy = accuracy_score([1 if label == -1 else 0 for label in true_labels], 
#                                          [1 if label == -1 else 0 for label in predicted_labels])

print("Overall accuracy:", accuracy)
print("Class 1 accuracy:", class_1_accuracy)
print("Class 0 accuracy:", class_0_accuracy)
# print("Class -1 accuracy:", class_minus_1_accuracy)
