==> GPCRdb_VEPs/3-1_from_GPCRdb_to_VESPA.ipynb <==
## README

make input files and command file(bash) for running VESPA

Input Files
Required: A single FASTA file containing all your wildtype sequences (note: this file can contain any number of sequences).

Optional: If you are only interested in a subset of possible mutations (specific mutations), you can add -m mutations.txt to the code line in Quickstart (note: per default all mutations are considered). Click here to head to the file format explanation of mutations.txt.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.metrics import r2_score
import re
import requests
import json


# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
uni_list = gene_list_raw['Uniprot'].tolist()
uni_list

gpcr_valid_missense = pd.read_csv('gpcr_valid_missense.csv')
gpcr_valid_missense

## Fetch FASTA file individually into a folder

def get_fasta(uniprot_id):
    # uniprot_id = "P07550"
    url = f"https://www.uniprot.org/uniprot/{uniprot_id}.fasta"

    response = requests.get(url)
    if response.ok:
        fasta_data = response.text
        with open(f"fasta/{uniprot_id}.fasta", "w") as f:
            f.write(fasta_data)
    else:
        print(f"Error retrieving FASTA sequence: {response.status_code} {response.reason}")

#fetch the sequence file
for i in range(len(uni_list)):
    get_fasta(uni_list[i])

## Fetch FASTA file altogether into a file

def get_fasta_merge(uniprot_id):
    # uniprot_id = "P07550"
    url = f"https://www.uniprot.org/uniprot/{uniprot_id}.fasta"

    response = requests.get(url)
    if response.ok:
        fasta_data = response.text
        with open("sequence.fasta", "a") as f:
            f.write(fasta_data)
    else:
        print(f"Error retrieving FASTA sequence: {response.status_code} {response.reason}")

# Load the array from the text file
unq_uniid_list = np.unique(gpcr_valid_missense['protein'].tolist())
unq_uniid_list

# put all fasta sequence into one fasta file
for i in range(len(unq_uniid_list)):
    get_fasta_merge(unq_uniid_list[i])
    

## Make mutation list for all target proteins into a txt file

VESPA! mutation position = position - 1!!!


for i in range(gpcr_valid_missense.shape[0]):
    # get the uniprot id according to entryname and convert the Series to a string with index excluded
    uniprot_id = gene_list_raw.loc[gene_list_raw['protein']==gpcr_valid_missense.iloc[i,4],'Uniprot'].to_string(index=False)
    position = gpcr_valid_missense.iloc[i,1]-1
    mut_info = str(gpcr_valid_missense.iloc[i,30])+str(position)+str(gpcr_valid_missense.iloc[i,31])
    with open("mutations.txt", 'a') as file:
        # write a new line to the file
        file.write(uniprot_id+"_"+mut_info+"\n")
    

fail to run vespa (which provides conservation scores and log-odds ratio) because the out of memory
only run VESPAI (conservation scores)

after getting the results, run the below parts:

## Process the output of VESPAI and output as csv file

import json
import pandas as pd
import os

# read in the map.json file
with open('output/map.json', 'r') as f:
    id_map = json.load(f)

# create an empty dataframe to store the combined data
combined_df = pd.DataFrame(columns=['Uniprot_ID', 'protein', 'Mutant', 'VESPAl'])

# loop over each file in the output folder
for filename in os.listdir('output'):
    if filename.endswith('.csv'):
        # get the uniprot id for this file from the filename
        id_num = int(filename.split('.')[0])
        uniprot_id = id_map[str(id_num)]

        protein_name = gene_list_raw.loc[gene_list_raw['Uniprot']== uniprot_id,'protein'].to_string(index=False)
        
        # read in the csv file and split the information on ';'
        csv_data = pd.read_csv(os.path.join('output', filename), delimiter=';')
        
        # add the uniprot id column and concatenate with the combined dataframe
        csv_data['Uniprot_ID'] = uniprot_id
        csv_data['protein'] = protein_name
        combined_df = pd.concat([combined_df, csv_data], ignore_index=True)

# print the resulting dataframe
print(combined_df)


# extract amino acid, position, and end amino acid from 'Mutant' column
combined_df[['mutation_from', 'mutation_pos', 'mutation_to']] = combined_df['Mutant'].str.extract(r'(\D)(\d+)(\D)')

# convert 'Position' column to integer type and plus 1 for further merge with GPCRdb data
combined_df['mutation_pos'] = combined_df['mutation_pos'].astype(int)+1
combined_df = combined_df.drop('Mutant', axis=1)
combined_df

combined_df.to_csv("gpcr_vespa_df.csv",index=False)
