==> GPCRdb_VEPs/4-2_FoldX_ddG_GPCRdb_merge.ipynb <==
# README
The purpose of this notebook is to process the output from FoldX (For variants from GPCRdb). the results are in different folders, we need to 
1: for each folder, read in the file starts with 'Dif_'(this contains the delta delta G) and the file starts with 'individual_list'(this contains the correspondence information between mutation and ddG) 
2: merge the mutation info and its corresponding ddG 
3: generate a dataframe and merge with Genebass data
4: output as csv file for further performance calculation

# Import packages

import pandas as pd
import numpy as np
import os
import seaborn as sns

import matplotlib.pyplot as plt

# Processing

# diff_energy = pd.read_csv('/Users/joannahench/Downloads/foldx5MacStd.tar_/GPCRdb/P0DMS8/Dif_P0DMS8.fxout',skiprows=8,sep='\t')
# diff_energy

# define the directory containing the protein folders
directory = '/Users/joannahench/Downloads/foldx5MacStd.tar_/GPCRdb'

# create an empty list to store the final data
final_data = []

# loop over each protein folder
for protein_folder in os.listdir(directory):

    # check if the item in the directory is a folder
    if os.path.isdir(os.path.join(directory, protein_folder)):


        # read in the Dif_ file as a csv, skipping the first 8 rows
        dif_file = pd.read_csv(os.path.join(directory, protein_folder, 'Dif_'+str(protein_folder)+'.fxout'), skiprows=8,sep = '\t')
        

        # read in the individual_list file as a csv with semicolon separator
        individual_list_file = pd.read_csv(os.path.join(directory, protein_folder, 'individual_list_'+str(protein_folder)+'.txt'), sep=';',header=None)


        # create a new dataframe with three columns: protein, total energy, and individual data
        new_data = pd.DataFrame({'Uniprot': protein_folder, 'FoldX': dif_file['total energy'], 'mutation data': individual_list_file[0]})

        # append the new data to the final data list
        final_data.append(new_data)


# concatenate all the new dataframes into a single dataframe
final_dataframe = pd.concat(final_data)

# split the 'mutation data' into 'mutation_from','mutation_to','mutation_pos'
final_dataframe['mutation_from'] = final_dataframe['mutation data'].str[0]
final_dataframe['mutation_to'] = final_dataframe['mutation data'].str[-1]
final_dataframe['mutation_pos'] = final_dataframe['mutation data'].str[2:-1].astype(int)


final_dataframe.reset_index(drop=True)
final_dataframe

# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
gene_list_raw

final_dataframe = final_dataframe.merge(gene_list_raw[['protein','gene','Uniprot']],on='Uniprot')
final_dataframe = final_dataframe.rename(columns={'gene':'SYMBOL'})
final_dataframe

# read in GPCRdb data
gpcr_missense = pd.read_csv('gpcr_filtered_mutants_0217.csv')
gpcr_missense

fx_df = pd.merge(gpcr_missense,final_dataframe,on=['mutation_pos','mutation_from','mutation_to','protein'])
fx_df

fx_df.to_csv('GPCR_FoldX_df.csv',index=False)
