==> GPCRdb_VEPs/5-1_RaSP_ddG_GPCRdb_merge.ipynb <==
# Import packages

import pandas as pd
import numpy as np
import os
import seaborn as sns

import matplotlib.pyplot as plt

# processing

# the directory containing the rasp predictions
directory = '/Users/joannahench/Documents/UCPH/MSc_Thesis/RaSP_pred'

# create an empty list to store the final data
rasp_data = pd.DataFrame()
# loop over each protein folder
for protein_file in os.listdir(directory):
    # read in the prediction output
    dif_file = pd.read_csv(os.path.join(directory, protein_file))
    rasp_data = pd.concat([rasp_data,dif_file])


rasp_data = rasp_data.rename(columns={'uniprot':'Uniprot','score_ml':'RaSP'})
rasp_data

# map protein names and gene names
# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
gene_list_raw.rename(columns={'gene':'SYMBOL'},inplace=True)
gene_list_raw

rasp_data_m = pd.merge(rasp_data,gene_list_raw,on=['Uniprot'])
rasp_data_m


rasp_data_m['mutation_from'] = rasp_data_m['variant'].str[0]
rasp_data_m['mutation_pos'] = rasp_data_m['variant'].str[1:-1].astype(int)
rasp_data_m['mutation_to'] = rasp_data_m['variant'].str[-1]
rasp_data_m

# read in GPCRdb data
gpcr_missense = pd.read_csv('gpcr_filtered_mutants_0217.csv')
gpcr_missense

rasp_df = pd.merge(gpcr_missense,rasp_data_m,on=['mutation_pos','mutation_from','mutation_to','protein'])
rasp_df

rasp_df.to_csv('GPCR_RaSP_df.csv',index=False)
